+++ app/web-app/swagger/v0/changelog.md
@@ -2,6 +2,10 @@
 
 Source: ${grailsApplication.config.grails.serverURL} - Version: ${de.laser.api.v0.ApiManager.VERSION}
 
+#### 0.113
+
+- internal refactoring
+
 #### 0.112
 
 - internal refactoring
@@ -12,7 +16,8 @@ Source: ${grailsApplication.config.grails.serverURL} - Version: ${de.laser.api.v
 
 #### 0.110
 
-- internal refactoring
+- added new attribute: `CostItem.isVisibleForSubscriber`
+- modified attribute: `CostItem.finalCostRounding`
 
 #### 0.109